* fix Utils sequence functions : use uint16_t (was uint32_t)

+++ NetworkLib/src/Utils.hpp
@@ -11,8 +11,8 @@ namespace Bousk
 	{
 		inline std::chrono::milliseconds Now();
 
-		inline bool IsSequenceNewer(const uint32_t sNew, const uint32_t sLast);
-		inline uint32_t SequenceDiff(const uint32_t sNew, const uint32_t sLast);
+		inline bool IsSequenceNewer(const uint16_t sNew, const uint16_t sLast);
+		inline uint16_t SequenceDiff(const uint16_t sNew, const uint16_t sLast);
 
 		inline void SetBit(uint64_t& bitfield, const uint8_t n);
 		inline void UnsetBit(uint64_t& bitfield, const uint8_t n);
